Replace blockDownloader with implementation in Fetcher
